Place executables in the same folder
Same behavior as with the premake4 buildsystem.

We only do this because `./lassie` uses `system()` to call
`./cmod`. This is a temporary fix; the true solution would be to
directly call into CMOD's code (since we already link with it!) rather
than using `system()`.

Once we stop using `system()` to call into CMOD, we can undo this
patch.