Fix segfault by zero-initializing all member variables
On startup, an "empty" project is added to the MainWindow, but the
constructor that gets called does not initialize all the
fields (specifically, eventAttributesView) of class
ProjectViewController. This meant that the program would occasionally
segfault on shutdown.

Set the default values of all pointer member variables to `nullptr`,
and make sure that no other variables are uninitialized on
construction.